Add an option to build vuk with system packages instead of the Vulkan SDK
`find_package(Vulkan)` fails if the Vulkan SDK is not installed on the system. However, on some environments (Linux, MSYS2) it's possible to install all the required Vulkan libraries through their package manager, keeping all the development files within the standard filesystem structure. The option `VUK_USE_VULKAN_SDK` is added, and enabled by default. Disabling it switches to a new CMake codepath that links all required libraries manually, all libraries are linked statically as per the project's convention. Additionally, with `VUK_LINK_TO_LOADER` on, the Vulkan loader is linked dynamically which seems to be strongly preferred by upstream.

New option tested on MSYS2, with `VUK_LINK_TO_LOADER` both on and off. Feel free to restructure the changes, or reject if you're not interested in supporting this configuration.